requirements.txt
factory-boy==3.0.1
Faker==4.1.2
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
